cleanup misc go and proto
